requirements.txt
asgiref==3.6.0
dj-database-url==1.3.0
Django==4.1.7
django-bootstrap-form==3.4
django-filter==22.1
django-widget-tweaks==1.4.12
gunicorn==20.1.0
psycopg2-binary==2.9.5
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
whitenoise==6.4.0
